+++ fabric/src/example/java/earth/terrarium/example/examples/ShaderExample.java
@@ -4,7 +4,9 @@ import com.teamresourceful.resourcefullib.common.color.Color;
 import earth.terrarium.example.base.ExampleScreen;
 import earth.terrarium.example.base.OlympusExample;
 import earth.terrarium.olympus.client.pipelines.RoundedRectangle;
+import earth.terrarium.olympus.client.pipelines.RoundedTexture;
 import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
 import org.jetbrains.annotations.NotNull;
 
@@ -18,31 +20,34 @@ public class ShaderExample extends ExampleScreen {
         graphics.pose().pushMatrix();
         graphics.pose().translate(Mth.cos(((System.currentTimeMillis() % 60000) / 1000f)) * 30f, 0);
 
+        // ARGB
         RoundedRectangle.draw(
                 graphics,
-                this.width / 2 - 150, 50,
+                this.width / 3 - this.width / 6 - 25, 50,
                 100, 100,
-                0x80000000, 0xFF0000FF,
-                15f, 2
+                0x80000000,
+                0xFF0000FF, 0xFFFF0000,
+                0xFF00FF00, 0xFF731f8f,
+                15, 2
         );
 
         RoundedRectangle.draw(
                 graphics,
-                this.width / 2 + 50, 50,
+                this.width * 2/3 - this.width / 6 - 25, 50,
                 100, 100,
                 Color.RAINBOW.withAlpha(0x80).getValue(), 0xFF00FF00,
                 15f, 2
         );
 
 
-//        RoundedTexture.draw(
-//                graphics,
-//                this.width / 2 + 50, 50,
-//                100, 100,
-//                ResourceLocation.fromNamespaceAndPath("olympus", "textures/gui/sprites/modal/modal.png"),
-//                0f, 0f, 1f, 1f,
-//                15f
-//        );
+        RoundedTexture.draw(
+                graphics,
+                this.width - this.width / 6 - 25, 50,
+                100, 100,
+                ResourceLocation.fromNamespaceAndPath("olympus", "textures/gui/sprites/modal/modal.png"),
+                0f, 0f, 1f, 1f,
+                15f
+        );
 
         graphics.pose().popMatrix();
     }