+++ changelog.md
@@ -1,1 +1,1 @@
-- Workaround issue where bounds arnt checked relatively in pips in vanilla
+-  Add ability to set color of each corner in rounded rect - J10a1n15